Fixes error when adding deposit.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3132 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/Accounts.cs
@@ -206,7 +206,7 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary></summary>
-		public static int[] GetDepositAccounts(){
+		public static long[] GetDepositAccounts(){
 			//No need to check RemotingRole; no call to db.
 			string depStr=PrefC.GetString(PrefName.AccountingDepositAccounts);
 			string[] depStrArray=depStr.Split(new char[] { ',' });
@@ -217,7 +217,7 @@ namespace OpenDentBusiness{
 				}
 				depAL.Add(PIn.PLong(depStrArray[i]));
 			}
-			int[] retVal=new int[depAL.Count];
+			long[] retVal=new long[depAL.Count];
 			depAL.CopyTo(retVal);
 			return retVal;
 		}
